Refactor: unified run script, separate agent configs with tyro
