+++ config.py
@@ -1,0 +1,69 @@
+import os
+from flax.struct import dataclass
+from typing import Union, Optional, Tuple
+from src.crl import CRL
+from src.ppo import PPO
+from src.sac import SAC
+from src.td3 import TD3
+
+# agent configurations
+AgentConfig = Union[PPO, SAC, TD3, CRL]
+
+
+@dataclass
+class RunConfig:
+    """General run configs
+    Args:
+        seed: random seed
+        env_name: the name of the environment to use
+        total_env_steps: the total number of environment steps to run
+        episode_length: the maximum length of an episode
+            NOTE: `num_envs * (episode_length - 1)` must be divisible by
+            `batch_size` due to the way data is stored in replay buffer.
+        num_envs: the number of parallel environments to use for rollouts
+            NOTE: `num_envs` must be divisible by the total number of chips since each
+            chip gets `num_envs // total_number_of_chips` environments to roll out
+            NOTE: `batch_size * num_minibatches` must be divisible by `num_envs` since
+            data generated by `num_envs` parallel envs gets used for gradient
+            updates over `num_minibatches` of data, where each minibatch has a
+            leading dimension of `batch_size`
+        max_devices_per_host: maximum number of chips to use per host process
+        num_eval_envs: the number of envs to use for evluation. Each env will run 1
+        episode, and all envs run in parallel during eval.
+        action_repeat: the number of timesteps to repeat an action
+        exp_name: the name of the experiment for logging
+        num_evals: the number of evals to run during the entire training run.
+            Increasing the number of evals increases total training time
+    """
+
+    env_name: str = "ant"
+    total_env_steps: int = 1_000_000
+    episode_length: int = 1001
+    eval_env: Optional[str] = None
+    num_envs: int = 128
+    num_eval_envs: int = 128
+    action_repeat: int = 1
+    num_evals: int = 50
+
+    seed: int = 0
+    exp_name: str = os.path.basename(__file__)[: -len(".py")]
+    backend: Optional[str] = None
+    log_wandb: bool = True
+    wandb_project_name: str = "jaxgcrl"
+    # online or offline
+    wandb_mode: str = "online"
+    wandb_group: str = "."
+    checkpoint: bool = False
+    visualization_interval: int = 5
+    vis_length: int = 1000
+    checkpoint_logdir: Optional[str] = None
+    max_devices_per_host: int = 1
+    cuda: bool = True
+
+
+@dataclass
+class Config:
+    # agent type
+    agent: AgentConfig
+    # run config
+    run: RunConfig
