+++ envs/ant_ball.py
@@ -1,12 +1,12 @@
 import os
 from typing import Tuple
-from brax import base
-from brax import math
+
+import jax
+import mujoco
+from brax import base, math
 from brax.envs.base import PipelineEnv, State
 from brax.io import mjcf
-import jax
 from jax import numpy as jnp
-import mujoco
 
 # This is based on original Ant environment from Brax
 # https://github.com/google/brax/blob/main/brax/envs/ant.py
