+++ envs/half_cheetah.py
@@ -1,9 +1,10 @@
 import os
 from typing import Tuple
+
+import jax
 from brax import base
 from brax.envs.base import PipelineEnv, State
 from brax.io import mjcf
-import jax
 from jax import numpy as jnp
 
 # This is based on original Half Cheetah environment from Brax
