+++ envs/humanoid.py
@@ -1,11 +1,11 @@
-from brax import actuator
-from brax import base
+import os
+
+import jax
+import mujoco
+from brax import actuator, base
 from brax.envs.base import PipelineEnv, State
 from brax.io import mjcf
-import jax
 from jax import numpy as jnp
-import mujoco
-import os
 
 # This is based on original Humanoid environment from Brax
 # https://github.com/google/brax/blob/main/brax/envs/humanoid.py
