+++ envs/pusher.py
@@ -1,9 +1,8 @@
-from brax import base
-from brax import math
+import jax
+from brax import base, math
 from brax.envs.base import PipelineEnv, State
 from brax.io import mjcf
 from etils import epath
-import jax
 from jax import numpy as jnp
 
 # This is based on original Pusher environment from Brax
