--- src/train.py
@@ -1,563 +1,0 @@
-import os
-import jax
-import flax
-import tyro
-import time
-import optax
-import wandb
-import pickle
-import random
-import wandb_osh
-import numpy as np
-import flax.linen as nn
-import jax.numpy as jnp
-
-from brax import envs
-from etils import epath
-from dataclasses import dataclass
-from typing import NamedTuple, Any, Optional
-from wandb_osh.hooks import TriggerWandbSyncHook
-from flax.training.train_state import TrainState
-from flax.linen.initializers import variance_scaling
-
-from envs.wrappers import TrajectoryIdWrapper
-from src.evaluator import CrlEvaluator
-from src.replay_buffer import TrajectoryUniformSamplingQueue
-from utils import Args, create_env, render_policy
-
-
-class Encoder(nn.Module):
-    repr_dim: int = 64
-    network_width: int = 256
-    network_depth: int = 4
-    skip_connections: int = 0  # 0 for no skip connections, >= 0 means the frequency of skip connections (every X layers)
-    use_relu: bool = False
-    use_ln: bool = False
-
-    @nn.compact
-    def __call__(self, data: jnp.ndarray):
-
-        lecun_unfirom = variance_scaling(1 / 3, "fan_in", "uniform")
-        bias_init = nn.initializers.zeros
-
-        if self.use_ln:
-            normalize = lambda x: nn.LayerNorm()(x)
-        else:
-            normalize = lambda x: x
-
-        if self.use_relu:
-            activation = nn.relu
-        else:
-            activation = nn.swish
-
-        x = data
-        for i in range(self.network_depth):
-            x = nn.Dense(self.network_width, kernel_init=lecun_unfirom, bias_init=bias_init)(x)
-            x = normalize(x)
-            x = activation(x)
-
-            if self.skip_connections:
-                if i == 0:
-                    skip = x
-                if i > 0 and i % self.skip_connections == 0:
-                    x = x + skip
-                    skip = x
-
-        x = nn.Dense(self.repr_dim, kernel_init=lecun_unfirom, bias_init=bias_init)(x)
-        return x
-
-class Actor(nn.Module):
-    action_size: int
-    network_width: int = 256
-    network_depth: int = 4
-    skip_connections: int = 0  # 0 for no skip connections, >= 0 means the frequency of skip connections (every X layers)
-    use_relu: bool = False
-    use_ln: bool = False
-    LOG_STD_MAX = 2
-    LOG_STD_MIN = -5
-
-    @nn.compact
-    def __call__(self, x):
-        if self.use_ln:
-            normalize = lambda x: nn.LayerNorm()(x)
-        else:
-            normalize = lambda x: x
-
-        if self.use_relu:
-            activation = nn.relu
-        else:
-            activation = nn.swish
-
-        lecun_unfirom = variance_scaling(1 / 3, "fan_in", "uniform")
-        bias_init = nn.initializers.zeros
-
-        print(f"x.shape: {x.shape}", flush=True)
-        for i in range(self.network_depth):
-            x = nn.Dense(self.network_width, kernel_init=lecun_unfirom, bias_init=bias_init)(x)
-            x = normalize(x)
-            x = activation(x)
-
-            if self.skip_connections:
-                if i == 0:
-                    skip = x
-                if i > 0 and i % self.skip_connections == 0:
-                    x = x + skip
-                    skip = x
-
-        mean = nn.Dense(self.action_size, kernel_init=lecun_unfirom, bias_init=bias_init)(x)
-        log_std = nn.Dense(self.action_size, kernel_init=lecun_unfirom, bias_init=bias_init)(x)
-
-        log_std = nn.tanh(log_std)
-        log_std = self.LOG_STD_MIN + 0.5 * (self.LOG_STD_MAX - self.LOG_STD_MIN) * (
-                    log_std + 1)  # From SpinUp / Denis Yarats
-
-        return mean, log_std
-
-
-@flax.struct.dataclass
-class TrainingState:
-    """Contains training state for the learner"""
-    env_steps: jnp.ndarray
-    gradient_steps: jnp.ndarray
-    actor_state: TrainState
-    critic_state: TrainState
-    alpha_state: TrainState
-
-class Transition(NamedTuple):
-    """Container for a transition"""
-    observation: jnp.ndarray
-    action: jnp.ndarray
-    reward: jnp.ndarray
-    discount: jnp.ndarray
-    extras: jnp.ndarray = ()  
-
-def load_params(path: str):
-    with epath.Path(path).open('rb') as fin:
-        buf = fin.read()
-    return pickle.loads(buf)
-
-def save_params(path: str, params: Any):
-    """Saves parameters in flax format."""
-    with epath.Path(path).open('wb') as fout:
-        fout.write(pickle.dumps(params))
-                   
-
-def train(args: Args):
-
-    args.env_steps_per_actor_step = args.num_envs * args.unroll_length
-    args.num_prefill_env_steps = args.min_replay_size * args.num_envs
-    args.num_prefill_actor_steps = np.ceil(args.min_replay_size / args.unroll_length)
-    args.num_training_steps_per_epoch = (args.total_env_steps - args.num_prefill_env_steps) // (args.num_evals * args.env_steps_per_actor_step)
-
-    run_name = f"{args.env_name}__{args.exp_name}__{args.seed}__{int(time.time())}"
-
-    if args.log_wandb:
-
-        if args.wandb_group ==  '.':
-            args.wandb_group = None
-            
-        wandb.init(
-            project=args.wandb_project_name,
-            mode=args.wandb_mode,
-            group=args.wandb_group,
-            config=vars(args),
-            name=run_name,
-            monitor_gym=True,
-            save_code=True,
-        )
-
-        if args.wandb_mode == 'offline':
-            wandb_osh.set_log_level("ERROR")
-            trigger_sync = TriggerWandbSyncHook()
-
-    random.seed(args.seed)
-    np.random.seed(args.seed)
-    key = jax.random.PRNGKey(args.seed)
-    key, buffer_key, env_key, eval_env_key, actor_key, sa_key, g_key = jax.random.split(key, 7)
-
-    # Environment setup    
-    env = create_env(args.env_name)
-    env = TrajectoryIdWrapper(env)
-    env = envs.training.wrap(
-        env,
-        episode_length=args.episode_length,
-        action_repeat=args.action_repeat,
-    )
-
-    if not args.eval_env:
-        eval_env_name = args.env_name
-    else:
-        eval_env_name = args.eval_env
-    eval_env = create_env(eval_env_name)
-    eval_env = TrajectoryIdWrapper(eval_env)
-    eval_env = envs.training.wrap(
-        eval_env,
-        episode_length=args.episode_length,
-        action_repeat=args.action_repeat,
-    )
-
-    # Dimensions definitions and sanity checks
-    action_size = env.action_size
-    state_size = env.state_dim
-    goal_size = len(env.goal_indices)
-    obs_size = state_size + goal_size 
-    assert obs_size == env.observation_size, f"obs_size: {obs_size}, env.observation_size: {env.observation_size}"
-
-    env_keys = jax.random.split(env_key, args.num_envs)
-    env_state = jax.jit(env.reset)(env_keys)
-    env.step = jax.jit(env.step)
-
-    # Network setup
-    # Actor
-    actor = Actor(action_size=action_size, network_width=args.h_dim, network_depth=args.n_hidden, skip_connections=args.skip_connections, use_relu=args.use_relu)
-    actor_state = TrainState.create(
-        apply_fn=actor.apply,
-        params=actor.init(actor_key, np.ones([1, obs_size])),
-        tx=optax.adam(learning_rate=args.policy_lr)
-    )
-
-    # Critic
-    sa_encoder = Encoder(repr_dim=args.repr_dim, network_width=args.h_dim, network_depth=args.n_hidden, skip_connections=args.skip_connections, use_relu=args.use_relu, use_ln=args.use_ln)
-    sa_encoder_params = sa_encoder.init(sa_key, np.ones([1, state_size+action_size]))
-    g_encoder = Encoder(repr_dim=args.repr_dim, network_width=args.h_dim, network_depth=args.n_hidden, skip_connections=args.skip_connections, use_relu=args.use_relu, use_ln=args.use_ln)
-    g_encoder_params = g_encoder.init(g_key, np.ones([1, goal_size]))
-    critic_state = TrainState.create(
-        apply_fn=None,
-        params={"sa_encoder": sa_encoder_params, "g_encoder": g_encoder_params},
-        tx=optax.adam(learning_rate=args.critic_lr),
-    )
-
-    # Entropy coefficient
-    target_entropy = -0.5 * action_size
-    log_alpha = jnp.asarray(0.0, dtype=jnp.float32)
-    alpha_state = TrainState.create(
-        apply_fn=None,
-        params={"log_alpha": log_alpha},
-        tx=optax.adam(learning_rate=args.alpha_lr),
-    )
-    
-    # Trainstate
-    training_state = TrainingState(
-        env_steps=jnp.zeros(()),
-        gradient_steps=jnp.zeros(()),
-        actor_state=actor_state,
-        critic_state=critic_state,
-        alpha_state=alpha_state,
-    )
-
-    #Replay Buffer
-    dummy_obs = jnp.zeros((obs_size,))
-    dummy_action = jnp.zeros((action_size,))
-
-    dummy_transition = Transition(
-        observation=dummy_obs,
-        action=dummy_action,
-        reward=0.0,
-        discount=0.0,
-        extras={
-            "state_extras": {
-                "truncation": 0.0,
-                "traj_id": 0.0,
-            }        
-        },
-    )
-
-    def jit_wrap(buffer):
-        buffer.insert_internal = jax.jit(buffer.insert_internal)
-        buffer.sample_internal = jax.jit(buffer.sample_internal)
-        return buffer
-    
-    replay_buffer = jit_wrap(
-            TrajectoryUniformSamplingQueue(
-                max_replay_size=args.max_replay_size,
-                dummy_data_sample=dummy_transition,
-                sample_batch_size=args.batch_size,
-                num_envs=args.num_envs,
-                episode_length=args.episode_length,
-            )
-        )
-    buffer_state = jax.jit(replay_buffer.init)(buffer_key)
-
-    def deterministic_actor_step(training_state, env, env_state, extra_fields):        
-        means, _ = actor.apply(training_state.actor_state.params, env_state.obs)
-        actions = nn.tanh( means )
-
-        nstate = env.step(env_state, actions)
-        state_extras = {x: nstate.info[x] for x in extra_fields}
-        
-        return nstate, Transition(
-            observation=env_state.obs,
-            action=actions,
-            reward=nstate.reward,
-            discount=1-nstate.done,
-            extras={"state_extras": state_extras},
-        )
-    
-    def actor_step(actor_state, env, env_state, key, extra_fields):        
-        means, log_stds = actor.apply(actor_state.params, env_state.obs)
-        stds = jnp.exp(log_stds)
-        actions = nn.tanh( means + stds * jax.random.normal(key, shape=means.shape, dtype=means.dtype) )
-
-        nstate = env.step(env_state, actions)
-        state_extras = {x: nstate.info[x] for x in extra_fields}
-        
-        return nstate, Transition(
-            observation=env_state.obs,
-            action=actions,
-            reward=nstate.reward,
-            discount=1-nstate.done,
-            extras={"state_extras": state_extras},
-        )
-
-    @jax.jit
-    def get_experience(actor_state, env_state, buffer_state, key):
-        @jax.jit
-        def f(carry, unused_t):
-            env_state, current_key = carry
-            current_key, next_key = jax.random.split(current_key)
-            env_state, transition = actor_step(actor_state, env, env_state, current_key, extra_fields=("truncation", "traj_id"))
-            return (env_state, next_key), transition
-
-        (env_state, _), data = jax.lax.scan(f, (env_state, key), (), length=args.unroll_length)
-
-        buffer_state = replay_buffer.insert(buffer_state, data)
-        return env_state, buffer_state
-
-    def prefill_replay_buffer(training_state, env_state, buffer_state, key):
-        @jax.jit
-        def f(carry, unused):
-            del unused
-            training_state, env_state, buffer_state, key = carry
-            key, new_key = jax.random.split(key)
-            env_state, buffer_state = get_experience(
-                training_state.actor_state,
-                env_state,
-                buffer_state,
-                key,
-            
-            )
-            training_state = training_state.replace(
-                env_steps=training_state.env_steps + args.env_steps_per_actor_step,
-            )
-            return (training_state, env_state, buffer_state, new_key), ()
-
-        return jax.lax.scan(f, (training_state, env_state, buffer_state, key), (), length=args.num_prefill_actor_steps)[0]
-
-    @jax.jit
-    def update_actor_and_alpha(transitions, training_state, key):
-        def actor_loss(actor_params, critic_params, log_alpha, transitions, key):
-            obs = transitions.observation           # expected_shape = batch_size, obs_size + goal_size
-            state = obs[:, :state_size]
-            future_state = transitions.extras["future_state"]
-            goal = future_state[:, env.goal_indices]
-            observation = jnp.concatenate([state, goal], axis=1)
-
-            means, log_stds = actor.apply(actor_params, observation)
-            stds = jnp.exp(log_stds)
-            x_ts = means + stds * jax.random.normal(key, shape=means.shape, dtype=means.dtype)
-            action = nn.tanh(x_ts)
-            log_prob = jax.scipy.stats.norm.logpdf(x_ts, loc=means, scale=stds)
-            log_prob -= jnp.log((1 - jnp.square(action)) + 1e-6)
-            log_prob = log_prob.sum(-1)           # dimension = B
-
-            sa_encoder_params, g_encoder_params = critic_params["sa_encoder"], critic_params["g_encoder"]
-            sa_repr = sa_encoder.apply(sa_encoder_params, jnp.concatenate([state, action], axis=-1))
-            g_repr = g_encoder.apply(g_encoder_params, goal)
-
-            qf_pi = -jnp.sqrt(jnp.sum((sa_repr - g_repr) ** 2, axis=-1))
-
-            actor_loss = jnp.mean( jnp.exp(log_alpha) * log_prob - (qf_pi) )
-
-            return actor_loss, log_prob
-
-        def alpha_loss(alpha_params, log_prob):
-            alpha = jnp.exp(alpha_params["log_alpha"])
-            alpha_loss = alpha * jnp.mean(jax.lax.stop_gradient(-log_prob - target_entropy))
-            return jnp.mean(alpha_loss)
-        
-        (actorloss, log_prob), actor_grad = jax.value_and_grad(actor_loss, has_aux=True)(training_state.actor_state.params, training_state.critic_state.params, training_state.alpha_state.params['log_alpha'], transitions, key)
-        new_actor_state = training_state.actor_state.apply_gradients(grads=actor_grad)
-
-        alphaloss, alpha_grad = jax.value_and_grad(alpha_loss)(training_state.alpha_state.params, log_prob)
-        new_alpha_state = training_state.alpha_state.apply_gradients(grads=alpha_grad)
-
-        training_state = training_state.replace(actor_state=new_actor_state, alpha_state=new_alpha_state)
-
-        metrics = {
-            "sample_entropy": -log_prob,
-            "actor_loss": actorloss,
-            "alph_aloss": alphaloss,   
-            "log_alpha": training_state.alpha_state.params["log_alpha"],
-        }
-
-        return training_state, metrics
-
-    @jax.jit
-    def update_critic(transitions, training_state, key):
-        def critic_loss(critic_params, transitions, key):
-            sa_encoder_params, g_encoder_params = critic_params["sa_encoder"], critic_params["g_encoder"]
-            
-            state = transitions.observation[:, :state_size]
-            action = transitions.action
-            
-            sa_repr = sa_encoder.apply(sa_encoder_params, jnp.concatenate([state, action], axis=-1))
-            g_repr = g_encoder.apply(g_encoder_params, transitions.observation[:, state_size:])
-            
-            # InfoNCE
-            logits = -jnp.sqrt(jnp.sum((sa_repr[:, None, :] - g_repr[None, :, :]) ** 2, axis=-1))       # shape = BxB
-            critic_loss = -jnp.mean(jnp.diag(logits) - jax.nn.logsumexp(logits, axis=1))
-
-            # logsumexp regularisation
-            logsumexp = jax.nn.logsumexp(logits + 1e-6, axis=1)
-            critic_loss += args.logsumexp_penalty_coeff * jnp.mean(logsumexp**2)
-
-            I = jnp.eye(logits.shape[0])
-            correct = jnp.argmax(logits, axis=1) == jnp.argmax(I, axis=1)
-            logits_pos = jnp.sum(logits * I) / jnp.sum(I)
-            logits_neg = jnp.sum(logits * (1 - I)) / jnp.sum(1 - I)
-
-            return critic_loss, (logsumexp, I, correct, logits_pos, logits_neg)
-            
-        (loss, (logsumexp, I, correct, logits_pos, logits_neg)), grad = jax.value_and_grad(critic_loss, has_aux=True)(training_state.critic_state.params, transitions, key)
-        new_critic_state = training_state.critic_state.apply_gradients(grads=grad)
-        training_state = training_state.replace(critic_state = new_critic_state)
-
-        metrics = {
-            "categorical_accuracy": jnp.mean(correct),
-            "logits_pos": logits_pos,
-            "logits_neg": logits_neg,
-            "logsumexp": logsumexp.mean(),
-            "critic_loss": loss,
-        }
-
-        return training_state, metrics
-    
-    @jax.jit
-    def sgd_step(carry, transitions):
-        training_state, key = carry
-        key, critic_key, actor_key, = jax.random.split(key, 3)
-
-        training_state, actor_metrics = update_actor_and_alpha(transitions, training_state, actor_key)
-        training_state, critic_metrics = update_critic(transitions, training_state, critic_key)
-        training_state = training_state.replace(gradient_steps = training_state.gradient_steps + 1)
-
-        metrics = {}
-        metrics.update(actor_metrics)
-        metrics.update(critic_metrics)
-        
-        return (training_state, key,), metrics
-
-    @jax.jit
-    def training_step(training_state, env_state, buffer_state, key):
-        experience_key1, experience_key2, sampling_key, training_key = jax.random.split(key, 4)
-
-        # update buffer
-        env_state, buffer_state = get_experience(
-            training_state.actor_state,
-            env_state,
-            buffer_state,
-            experience_key1,
-        )
-
-        training_state = training_state.replace(
-            env_steps=training_state.env_steps + args.env_steps_per_actor_step,
-        )
-
-        # sample actor-step worth of transitions
-        buffer_state, transitions = replay_buffer.sample(buffer_state)
-
-        # process transitions for training
-        batch_keys = jax.random.split(sampling_key, transitions.observation.shape[0])
-        transitions = jax.vmap(TrajectoryUniformSamplingQueue.flatten_crl_fn, in_axes=(None, 0, 0))((args.discounting, state_size, tuple(env.goal_indices)), transitions, batch_keys)  
-        transitions = jax.tree_util.tree_map(lambda x: jnp.reshape(x, (-1,) + x.shape[2:], order="F"),transitions)
-        
-        # permute transitions
-        permutation = jax.random.permutation(experience_key2, len(transitions.observation))
-        transitions = jax.tree_util.tree_map(lambda x: x[permutation], transitions)
-        transitions = jax.tree_util.tree_map(lambda x: jnp.reshape(x, (-1, args.batch_size) + x.shape[1:]),transitions)
-
-        # take actor-step worth of training-step
-        (training_state, _,), metrics = jax.lax.scan(sgd_step, (training_state, training_key), transitions)
-
-        return (training_state, env_state, buffer_state,), metrics
-
-    @jax.jit
-    def training_epoch(
-        training_state,
-        env_state,
-        buffer_state,
-        key,
-    ):  
-        @jax.jit
-        def f(carry, unused_t):
-            ts, es, bs, k = carry
-            k, train_key = jax.random.split(k, 2)
-            (ts, es, bs,), metrics = training_step(ts, es, bs, train_key)
-            return (ts, es, bs, k), metrics
-
-        (training_state, env_state, buffer_state, key), metrics = jax.lax.scan(f, (training_state, env_state, buffer_state, key), (), length=args.num_training_steps_per_epoch)
-        
-        metrics["buffer_current_size"] = replay_buffer.size(buffer_state)
-        return training_state, env_state, buffer_state, metrics
-
-    key, prefill_key = jax.random.split(key, 2)
-
-    training_state, env_state, buffer_state, _ = prefill_replay_buffer(
-        training_state, env_state, buffer_state, prefill_key
-    )
-
-    '''Setting up evaluator'''
-    evaluator = CrlEvaluator(
-        deterministic_actor_step,
-        eval_env,
-        num_eval_envs=args.num_eval_envs,
-        episode_length=args.episode_length,
-        key=eval_env_key,
-    )
-
-    training_walltime = 0
-    print('starting training....')
-    for ne in range(args.num_evals):
-        
-        t = time.time()
-
-        key, epoch_key = jax.random.split(key)
-        training_state, env_state, buffer_state, metrics = training_epoch(training_state, env_state, buffer_state, epoch_key)
-        
-        metrics = jax.tree_util.tree_map(jnp.mean, metrics)
-        metrics = jax.tree_util.tree_map(lambda x: x.block_until_ready(), metrics)
-
-        epoch_training_time = time.time() - t
-        training_walltime += epoch_training_time
-
-        sps = (args.env_steps_per_actor_step * args.num_training_steps_per_epoch) / epoch_training_time
-        metrics = {
-            "training/sps": sps,
-            "training/walltime": training_walltime,
-            "training/envsteps": training_state.env_steps.item(),
-            **{f"training/{name}": value for name, value in metrics.items()},
-        }
-
-        metrics = evaluator.run_evaluation(training_state, metrics)
-
-        print(metrics)
-        if ne % args.visualization_interval == 0:
-            render_policy(training_state, args.run_dir, env, actor, eval_env_name, args.vis_length)
-
-        if args.checkpoint:
-            # Save current policy and critic params.
-            params = (training_state.alpha_state.params, training_state.actor_state.params, training_state.critic_state.params)
-            path = f"{args.ckpt_dir}/step_{int(training_state.env_steps)}.pkl"
-            save_params(path, params)
-        
-        if args.log_wandb:
-            wandb.log(metrics, step=ne)
-
-            if args.wandb_mode == 'offline':
-                trigger_sync()
-    
-    if args.checkpoint:
-        # Save current policy and critic params.
-        params = (training_state.alpha_state.params, training_state.actor_state.params, training_state.critic_state.params)
-        path = f"{args.ckpt_dir}/final.pkl"
-        save_params(path, params)
