+++ utils.py
@@ -1,97 +1,36 @@
 import argparse
-from dataclasses import dataclass
+import math
 import os
 from collections import namedtuple
 from datetime import datetime
-from typing import List, Optional
+from typing import List
 
-import jax
-import math
-from brax.io import html
 import flax.linen as nn
+import jax
+import wandb_osh
+from brax.io import html
 from matplotlib import pyplot as plt
-import wandb
+from wandb_osh.hooks import TriggerWandbSyncHook
 
+import wandb
 from envs.ant import Ant
-from envs.half_cheetah import Halfcheetah
-from envs.reacher import Reacher
-from envs.pusher import Pusher, PusherReacher
-from envs.pusher2 import Pusher2
 from envs.ant_ball import AntBall
+from envs.ant_ball_maze import AntBallMaze
 from envs.ant_maze import AntMaze
+from envs.ant_push import AntPush
+from envs.half_cheetah import Halfcheetah
 from envs.humanoid import Humanoid
 from envs.humanoid_maze import HumanoidMaze
-from envs.ant_push import AntPush
-from envs.manipulation.arm_reach import ArmReach
+from envs.manipulation.arm_binpick_easy import ArmBinpickEasy
+from envs.manipulation.arm_binpick_hard import ArmBinpickHard
 from envs.manipulation.arm_grasp import ArmGrasp
 from envs.manipulation.arm_push_easy import ArmPushEasy
 from envs.manipulation.arm_push_hard import ArmPushHard
-from envs.manipulation.arm_binpick_easy import ArmBinpickEasy
-from envs.manipulation.arm_binpick_hard import ArmBinpickHard
-from envs.ant_ball_maze import AntBallMaze
+from envs.manipulation.arm_reach import ArmReach
+from envs.pusher import Pusher, PusherReacher
+from envs.pusher2 import Pusher2
+from envs.reacher import Reacher
 from envs.simple_maze import SimpleMaze
-
-
-@dataclass
-class Args:
-    exp_name: str = os.path.basename(__file__)[: -len(".py")]
-    seed: int = 1
-    cuda: bool = True
-    log_wandb: bool = True
-    wandb_project_name: str = "exploration"
-    wandb_mode: str = 'online'
-    wandb_group: str = '.'
-    checkpoint: bool = False
-    visualization_interval: int = 5
-    vis_length: int = 1000
-
-    # Environment specific arguments
-    env_name: str = "ant"
-    episode_length: int = 1000
-    backend: Optional[str] = None
-    eval_env: Optional[str] = None
-    action_repeat: int = 1
-    num_eval_envs: int = 128
-    use_dense_reward: bool = False
-
-    # Algorithm specific arguments
-    total_env_steps: int = 50000000
-    num_evals: int = 50
-    num_envs: int = 1024
-    policy_lr: float = 3e-4
-    critic_lr: float = 3e-4
-    alpha_lr: float = 3e-4
-    batch_size: int = 256
-    discounting: float = 0.99
-    logsumexp_penalty_coeff: float = 0.1
-    train_step_multiplier: int = 1
-    use_her: bool = False
-    disable_entropy_actor: bool = False
-
-    max_replay_size: int = 10000
-    min_replay_size: int = 1000
-    unroll_length: int  = 62
-    h_dim: int = 256
-    n_hidden: int = 2
-    skip_connections: int = 4
-    use_relu: bool = False
-    repr_dim: int = 64
-    use_ln: bool = False
-
-    # to be filled in runtime
-    env_steps_per_actor_step : int = 0
-    """number of env steps per actor step (computed in runtime)"""
-    num_prefill_env_steps : int = 0
-    """number of env steps to fill the buffer before starting training (computed in runtime)"""
-    num_prefill_actor_steps : int = 0
-    """number of actor steps to fill the buffer before starting training (computed in runtime)"""
-    num_training_steps_per_epoch : int = 0
-    """the number of training steps per epoch(computed in runtime)"""
-    run_dir: str = ''
-    """the directory to save the run"""
-    ckpt_dir: str = ''
-    """the directory to save the checkpoints"""
-    utd_ratio: float = 0
 
 
 def create_env(env_name: str, backend: str = None, **kwargs) -> object:
@@ -123,13 +62,17 @@ def create_env(env_name: str, backend: str = None, **kwargs) -> object:
         env = AntPush(backend=backend or "mjx")
     elif "maze" in env_name:
         if "ant_ball" in env_name:
-            env = AntBallMaze(backend=backend or "spring", maze_layout_name=env_name[9:])
+            env = AntBallMaze(
+                backend=backend or "spring", maze_layout_name=env_name[9:]
+            )
         elif "ant" in env_name:
             # Possible env_name = {'ant_u_maze', 'ant_big_maze', 'ant_hardest_maze'}
             env = AntMaze(backend=backend or "spring", maze_layout_name=env_name[4:])
         elif "humanoid" in env_name:
             # Possible env_name = {'humanoid_u_maze', 'humanoid_big_maze', 'humanoid_hardest_maze'}
-            env = HumanoidMaze(backend=backend or "spring", maze_layout_name=env_name[9:])
+            env = HumanoidMaze(
+                backend=backend or "spring", maze_layout_name=env_name[9:]
+            )
         else:
             # Possible env_name = {'simple_u_maze', 'simple_big_maze', 'simple_hardest_maze'}
             env = SimpleMaze(backend=backend or "spring", maze_layout_name=env_name[7:])
@@ -162,28 +105,6 @@ def create_env(env_name: str, backend: str = None, **kwargs) -> object:
     return env
 
 
-def create_eval_env(args: argparse.Namespace) -> object:
-    """
-    Creates an evaluation environment based on the provided arguments.
-
-    This function generates a new environment specifically for evaluation based on
-    args.eval_env. If args.eval_env is not specified, the function returns None.
-
-    Args:
-        args (argparse.Namespace): The arguments containing configuration for the
-                                   environment, including eval_env.
-
-    Returns:
-        object: The created evaluation environment, or None if args.eval_env is not
-                specified.
-    """
-    if not args.eval_env:
-        return None
-    
-    eval_arg = argparse.Namespace(**vars(args))
-    eval_arg.env_name = args.eval_env
-    return create_env(**vars(eval_arg))
-
 def get_env_config(args: argparse.Namespace):
     """
     Generate and validate environment configuration based on input arguments.
@@ -194,7 +115,7 @@ def get_env_config(args: argparse.Namespace):
 
     Parameters
     ----------
-    args : argparse.Namespace
+    args : dataclass
         The input arguments containing the environment name and other configuration settings.
 
     Returns
@@ -208,20 +129,37 @@ def get_env_config(args: argparse.Namespace):
         If the specified environment name is not in the list of legal environments or does not
         contain the word 'maze'.
     """
-    legal_envs = ["reacher", "cheetah", "pusher_easy", "pusher_hard", "pusher_reacher", "pusher2",
-                  "ant", "ant_push", "ant_ball", "humanoid", "arm_reach", "arm_grasp",
-                  "arm_push_easy", "arm_push_hard", "arm_binpick_easy", "arm_binpick_hard"]
+    legal_envs = [
+        "reacher",
+        "cheetah",
+        "pusher_easy",
+        "pusher_hard",
+        "pusher_reacher",
+        "pusher2",
+        "ant",
+        "ant_push",
+        "ant_ball",
+        "humanoid",
+        "arm_reach",
+        "arm_grasp",
+        "arm_push_easy",
+        "arm_push_hard",
+        "arm_binpick_easy",
+        "arm_binpick_hard",
+    ]
     if args.env_name not in legal_envs and "maze" not in args.env_name:
         raise ValueError(f"Unknown environment: {args.env_name}")
 
     # TODO: round num_envs to nearest valid value instead of throwing error
     if ((args.episode_length - 1) * args.num_envs) % args.batch_size != 0:
-        raise ValueError("(episode_length - 1) * num_envs must be divisible by batch_size")
+        raise ValueError(
+            "(episode_length - 1) * num_envs must be divisible by batch_size"
+        )
 
     args_dict = vars(args)
     Config = namedtuple("Config", [*args_dict.keys()])
     config = Config(*args_dict.values())
-    
+
     return config
 
 
@@ -236,7 +174,15 @@ class MetricsRecorder:
     exp_dir (str): Directory to save renders to.
     exp_name (str): Experiment name for naming rendered trajectory visualizations.
     """
-    def __init__(self, total_env_steps: int, metrics_to_collect: List[str], exp_dir, exp_name):
+
+    def __init__(
+        self,
+        total_env_steps: int,
+        metrics_to_collect: List[str],
+        exp_dir,
+        exp_name,
+        mode,
+    ):
         self.x_data = []
         self.y_data = {}
         self.y_data_err = {}
@@ -244,12 +190,17 @@ class MetricsRecorder:
         self.metrics_to_collect = metrics_to_collect
         self.exp_dir = exp_dir
         self.exp_name = exp_name
+        self.mode = mode
 
         self.max_x, self.min_x = total_env_steps * 1.1, 0
 
+        if mode == "offline":
+            wandb_osh.set_log_level("ERROR")
+        self.trigger_sync = TriggerWandbSyncHook()
+
     def record(self, num_steps, metrics):
         self.times.append(datetime.now())
-        self.x_data.append(num_steps)
+        self.x_data.append(int(num_steps))
 
         for key, value in metrics.items():
             if key not in self.y_data:
@@ -265,6 +216,9 @@ class MetricsRecorder:
             data_to_log[key] = value[-1]
         data_to_log["step"] = self.x_data[-1]
         wandb.log(data_to_log, step=self.x_data[-1])
+
+        if self.mode == "offline":
+            self.trigger_sync()
 
     def plot_progress(self):
         num_plots = len(self.y_data)
@@ -292,23 +246,32 @@ class MetricsRecorder:
 
     def print_progress(self):
         for idx, (key, y_values) in enumerate(self.y_data.items()):
-            print(f"step: {self.x_data[-1]}, {key}: {y_values[-1]:.3f} +/- {self.y_data_err[key][-1]:.3f}")
+            print(
+                f"step: {self.x_data[-1]}, {key}: {y_values[-1]:.3f} +/- {self.y_data_err[key][-1]:.3f}"
+            )
 
     def print_times(self):
         print(f"time to jit: {self.times[1] - self.times[0]}")
         print(f"time to train: {self.times[-1] - self.times[1]}")
-        
+
     def progress(self, num_steps, metrics, make_policy, params, env, do_render=True):
         for key in self.metrics_to_collect:
             self.ensure_metric(metrics, key)
-        
+
         if do_render:
             render(make_policy, params, env, self.exp_dir, self.exp_name, num_steps)
-        
-        self.record(num_steps, {key: value for key, value in metrics.items() if key in self.metrics_to_collect})
+
+        self.record(
+            num_steps,
+            {
+                key: value
+                for key, value in metrics.items()
+                if key in self.metrics_to_collect
+            },
+        )
         self.log_wandb()
         self.print_progress()
-    
+
     @staticmethod
     def ensure_metric(metrics, key):
         if key not in metrics:
@@ -316,6 +279,7 @@ class MetricsRecorder:
         else:
             if math.isnan(metrics[key]):
                 raise Exception(f"Metric: {key} is Nan")
+
 
 def render(make_policy, params, env, exp_dir, exp_name, num_steps):
     """
@@ -357,21 +321,24 @@ def render(make_policy, params, env, exp_dir, exp_name, num_steps):
     for i in range(5000):
         rollout.append(state.pipeline_state)
         key, subkey = jax.random.split(key)
-        action, _ = jit_policy(state.obs[None], subkey) # Policy requires batched dimension
-        action = action[0] # Remove batch dimension
+        action, _ = jit_policy(
+            state.obs[None], subkey
+        )  # Policy requires batched dimension
+        action = action[0]  # Remove batch dimension
         state = jit_env_step(state, action)
         if i % 1000 == 0:
             key, subkey = jax.random.split(key)
             state = jit_env_reset(rng=subkey)
 
-    url = html.render(env.sys.tree_replace({"opt.timestep": env.dt}), rollout, height=1024)
+    url = html.render(
+        env.sys.tree_replace({"opt.timestep": env.dt}), rollout, height=1024
+    )
     with open(os.path.join(exp_dir, f"{exp_name}_{num_steps}.html"), "w") as file:
         file.write(url)
     wandb.log({"render": wandb.Html(url)})
 
 
-
-def render_policy(training_state, save_path, env, actor, eval_env_name, vis_length):
+def render_policy(params, save_path, env, actor, eval_env, vis_length):
     """Renders the policy and saves it as an HTML file."""
 
     # JIT compile the rollout function
@@ -384,7 +351,7 @@ def render_policy(training_state, save_path, env, actor, eval_env_name, vis_leng
 
     rollout_states = []
     for i in range(10):
-        env = create_env(eval_env_name)
+        env = create_env(eval_env) if type(eval_env) == str else eval_env
 
         # Initialize environment
         rng = jax.random.PRNGKey(seed=i + 1)
@@ -392,7 +359,7 @@ def render_policy(training_state, save_path, env, actor, eval_env_name, vis_leng
 
         # Collect rollout using jitted function
         for _ in range(vis_length):
-            env_state, current_state = policy_step(env_state, training_state.actor_state.params)
+            env_state, current_state = policy_step(env_state, params)
             rollout_states.append(current_state.pipeline_state)
 
     # Render and save
